feat(Logic): add the shopping cart functionality.
- use context API to write the shopping cart logic and use it to handle various components that need to recieve this functions.
