+++ src/components/card/GadgetCard.tsx
@@ -1,11 +1,8 @@
 // Company Gadget Card Component
 import { styled } from "@stitches/react";
-import { useState } from "react";
-
 import { FilledButton } from "../buttons/FilledButton";
 import { CartButton } from "../buttons/CartButton";
 import { StateButton } from "../buttons/StateButton";
-import { CartModal } from "../modal/CartModal";
 
 export type GadgetCardProps = {
   GadgetName: string;
@@ -94,11 +91,6 @@ const SecondaryButtonContainer = styled("div", {
 });
 
 export const GadgetCard = (props: GadgetCardProps) => {
-  const [showModal, setShowModal] = useState(false);
-
-  const openModal = () => {
-    setShowModal((prev) => !prev);
-  };
   return (
     <GadgetCardContainer
       style={
@@ -130,9 +122,8 @@ export const GadgetCard = (props: GadgetCardProps) => {
           {props.hasSpec ? <PriceSpan>{props.GadgetPrice}</PriceSpan> : ""}
           {props.hasSpec ? (
             <SecondaryButtonContainer>
-              <CartButton text="Add to Cart" onclick={openModal} />
+              <CartButton id={props.id} text="Add to Cart" />
               <StateButton id={props.id} />
-              <CartModal showModal={showModal} setShowModal={setShowModal} />
             </SecondaryButtonContainer>
           ) : (
             <FilledButton text="See Product" />
